Move location manager to its own Locations nav section
Separates the Location Manager and Archive/Delete flows from the Settings
page into a dedicated /locations route, and adds a Locations nav item to
the admin sidebar between Bookings and Notifications.
